+++ src/app/App.tsx
@@ -1,19 +1,11 @@
 import { RouterProvider, ReactQueryProvider } from "./provider"
-import PostsManagerPage from "../pages/PostsManagerPage.tsx"
-import { Header } from "../shared/ui/header"
-import { Footer } from "../shared/ui/footer"
+import { PostsManagerPage } from "../pages/postsManager/ui"
 
 const App = () => {
   return (
     <ReactQueryProvider>
       <RouterProvider>
-        <div className="flex flex-col min-h-screen">
-          <Header />
-          <main className="flex-grow container mx-auto px-4 py-8">
-            <PostsManagerPage />
-          </main>
-          <Footer />
-        </div>
+        <PostsManagerPage />
       </RouterProvider>
     </ReactQueryProvider>
   )
