+++ src/pages/postsManager/ui/index.ts
@@ -1,0 +1,1 @@
+export { PostsManagerPage } from "./PostsManagerPage.tsx"
