+++ src/widgets/postsManager/ui/index.ts
@@ -1,0 +1,1 @@
+export { PostsManagerWidget } from "./PostsManagerWidget.tsx"
